Conflicts:
	example/oglplus/028_depth_map_sss.cpp
	include/oglplus/ext/NV_path_rendering/path.hpp
	include/oglplus/ext/NV_path_rendering/path_array.hpp

+++ example/oglplus/013_striped_cubes.cpp
@@ -4,7 +4,7 @@
  *
  *  @oglplus_screenshot{013_striped_cubes}
  *
- *  Copyright 2008-2014 Matus Chochlik. Distributed under the Boost
+ *  Copyright 2008-2015 Matus Chochlik. Distributed under the Boost
  *  Software License, Version 1.0. (See accompanying file
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  *
@@ -143,7 +143,7 @@ public:
 		projection_matrix.Set(
 			CamMatrixf::PerspectiveX(
 				Degrees(60),
-				double(width)/height,
+				float(width)/height,
 				1, 30
 			)
 		);
