+++ example/oglplus/022_parallax_map.cpp
@@ -251,7 +251,7 @@ public:
 		projection_matrix.Set(
 			CamMatrixf::PerspectiveX(
 				Degrees(54),
-				double(width)/height,
+				float(width)/height,
 				1, 10
 			)
 		);
