+++ example/oglplus/028_ripples.cpp
@@ -405,7 +405,7 @@ public:
 		gl.Viewport(width, height);
 		perspective = CamMatrixf::PerspectiveX(
 			Degrees(60),
-			double(width)/height,
+			float(width)/height,
 			1, 100
 		);
 	}
