+++ example/oglplus/031_motion_blur.cpp
@@ -4,7 +4,7 @@
  *
  *  @oglplus_screenshot{031_motion_blur}
  *
- *  Copyright 2008-2014 Matus Chochlik. Distributed under the Boost
+ *  Copyright 2008-2015 Matus Chochlik. Distributed under the Boost
  *  Software License, Version 1.0. (See accompanying file
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  *
@@ -548,7 +548,7 @@ public:
 		auto projection =
 			CamMatrixf::PerspectiveX(
 				Degrees(60),
-				double(width)/height,
+				float(width)/height,
 				1, 300
 			);
