+++ example/oglplus/040_jelly_cube.cpp
@@ -1491,7 +1491,7 @@ public:
 
 		auto perspective = CamMatrixf::PerspectiveX(
 			Degrees(70),
-			double(width)/height,
+			float(width)/height,
 			1, 200
 		);
